Progressed wiring in setup
I came to a very unfortunate realization that Chrome's memory restrictions have
capped the keyring to only a few contact entries. I will have to drastically
re-architect the design of the keyring. As a quick fix, though, I might just
load everything into memory upon launch. I'm too tired to actually explain this
properly. Is this really the place to be explaining it?