Experimented with different node modification
It's far more efficient to listen for clicks on the whole body rather than to
listen for new elements and adding listeners to each new element. This method
is also more flexible in case there isn't a class I can use to distinguish
encrypted nodes. I'd still like to keep the old new-element listener around in
order to change CSS or to decrypt in place.  # Please enter the commit message
for your changes. Lines starting